Реализована нативная часть для Android

+++ lib/battery.dart
@@ -11,7 +11,7 @@ class Battery {
 
   static const MethodChannel _channel = MethodChannel(_channelName);
   static bool _initialized = false;
-  static ValueChanged<double>? _onBatteryPowerChanged;
+  static ValueChanged<int>? _onBatteryPowerChanged;
 
   static Future<String?> get platformVersion async {
     final version = await _channel.invokeMethod<String?>('getPlatformVersion');
@@ -19,16 +19,16 @@ class Battery {
     return version;
   }
 
-  /// Получение текущего зараяда батарееи в процентах (0.0 .. 1.0).
-  static Future<double?> get power async =>
-      _channel.invokeMethod<double?>(_methodGetPower);
+  /// Получение текущего зараяда батарееи в процентах (0..100).
+  static Future<int?> get power async =>
+      _channel.invokeMethod<int?>(_methodGetPower);
 
   /// Колбэк на изменение заряда батареи.
-  static ValueChanged<double>? get onBatteryPowerChanged =>
+  static ValueChanged<int>? get onBatteryPowerChanged =>
       _onBatteryPowerChanged;
 
   /// Установка колбэка на изменение заряда батареи.
-  static set onBatteryPowerChanged(ValueChanged<double>? value) {
+  static set onBatteryPowerChanged(ValueChanged<int>? value) {
     _onBatteryPowerChanged = value;
     if (value != null) {
       if (!_initialized) _init();
@@ -42,7 +42,7 @@ class Battery {
     _channel.setMethodCallHandler((call) async {
       switch (call.method) {
         case _methodPowerChangeNotify:
-          _onBatteryPowerChanged?.call(call.arguments as double);
+          _onBatteryPowerChanged?.call(call.arguments as int);
       }
     });
     _initialized = true;
